update update skill event

+++ src/main/java/org/common/event/SkillAtom.java
@@ -11,6 +11,6 @@ import lombok.NoArgsConstructor;
 @Builder
 public class SkillAtom {
     private String name; // versapath atom name
-    private Long courseModuleId; // id for tracking progress
-    private Long pageId; // id for fetching content
+    private int courseModuleId; // id for tracking progress
+    private int pageId; // id for fetching content
 }
